requirements.txt
exceptiongroup==1.2.1
fastapi==0.110.3
gunicorn==22.0.0
huggingface-hub==0.22.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
scikit-learn==1.4.2
pandas==2.2.2
requests==2.31.0
sentence-transformers==2.7.0
tokenizers==0.19.1
transformers==4.40.1
uvicorn==0.29.0
oauth2client==4.1.3
oauthlib==3.2.2
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
gspread==6.1.2